learningpath-api+search-api: Move `MyNDLAApiClient` to shared lib
Moving `MyNDLAApiClient` to a shared place to maybe allow for sharing
code that fetches users when authenticating feide tokens.
