+++ search-api/src/test/scala/no/ndla/searchapi/service/search/LearningPathIndexServiceTest.scala
@@ -38,7 +38,7 @@ class LearningPathIndexServiceTest
   }
 
   override def beforeAll(): Unit = {
-    when(myndlaapiClient.getStatsFor(any, any)).thenReturn(Success(List.empty))
+    when(myndlaApiClient.getStatsFor(any, any)).thenReturn(Success(List.empty))
   }
 
   override val converterService       = new ConverterService
